Rename shared build logic & revamp dependency mgmt from Plugin side
