+++ plugin/android-junit5/src/test/resources/test-projects/settings.gradle.kts.template
@@ -10,6 +10,7 @@
 pluginManagement {
   repositories {
     google()
+    mavenCentral()
   }
   resolutionStrategy {
     eachPlugin {
