+++ plugin/settings.gradle.kts
@@ -1,3 +1,3 @@
 rootProject.name = "android-junit5-plugin"
-
+includeBuild("../build-logic")
 include(":android-junit5")
